+++ lib/services/safety/base_stateful.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nft/services/app/dynamic_size.dart';
+import 'package:nft/utils/app_extension.dart';
 import 'package:nft/utils/app_theme.dart';
 
 /// Remember call super.build(context) in widget
@@ -11,7 +12,7 @@ abstract class BaseStateful<T extends StatefulWidget> extends State<T>
   @mustCallSuper
   @protected
   void initDependencies(BuildContext context) {
-    appTheme = context.theme();
+    appTheme = context.appTheme();
   }
 
   @protected