+++ lib/services/safety/base_stateless.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nft/services/app/dynamic_size.dart';
+import 'package:nft/utils/app_extension.dart';
 import 'package:nft/utils/app_theme.dart';
 
 /// Remember call super.build(context) in widget
@@ -13,7 +14,7 @@ abstract class BaseStateless extends StatelessWidget with DynamicSize {
   @mustCallSuper
   @protected
   void initDependencies(BuildContext context) {
-    appTheme = context.theme();
+    appTheme = context.appTheme();
   }
 
   @protected