+++ lib/services/safety/page_stateful.dart
@@ -6,7 +6,7 @@ import 'package:nft/services/app/locale_provider.dart';
 import 'package:nft/services/rest_api/api_error.dart';
 import 'package:nft/services/rest_api/api_error_type.dart';
 import 'package:nft/services/safety/base_stateful.dart';
-import 'package:nft/utils/app_route.dart';
+import 'package:nft/utils/app_extension.dart';
 import 'package:provider/provider.dart';
 
 abstract class PageStateful<T extends StatefulWidget> extends BaseStateful<T>
@@ -23,13 +23,16 @@ abstract class PageStateful<T extends StatefulWidget> extends BaseStateful<T>
   }
 
   @override
-  Future<void> onApiError(dynamic error) async {
+  Future<int> onApiError(dynamic error) async {
     final ApiErrorType errorType = parseApiErrorType(error);
-    await AppDialogProvider.show(context, errorType.message, title: 'Error');
-    await Future<void>.delayed(const Duration(seconds: 1));
-    if (errorType.code == ApiErrorCode.unauthorized) {
-      logout(context);
+    if (errorType.message != null && errorType.message.isNotEmpty) {
+      await AppDialogProvider.show(context, errorType.message, title: 'Error');
     }
+    if (errorType.code == ApiErrorCode.unauthorized) {
+      await logout(context);
+      return 1;
+    }
+    return 0;
   }
 
   /// Logout function