+++ lib/widgets/p_appbar_empty.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:nft/utils/app_extension.dart';
 import 'package:nft/utils/app_theme.dart';
 import 'package:nft/widgets/p_material.dart';
 
@@ -10,7 +11,7 @@ class PAppBarEmpty extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final AppTheme theme = context.theme();
+    final AppTheme theme = context.appTheme();
     return PMaterial(
       child: Scaffold(
         backgroundColor: theme.backgroundColor ?? Colors.transparent,