+++ lib/widgets/p_appbar_transparency.dart
@@ -1,19 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:nft/utils/app_extension.dart';
 import 'package:nft/utils/app_theme.dart';
 import 'package:nft/widgets/p_material.dart';
 
 class PAppBarTransparency extends StatelessWidget {
   const PAppBarTransparency(
-      {@required this.child, this.forceStatusIconLight, Key key})
+      {this.body, this.child, this.forceStatusIconLight, Key key})
       : super(key: key);
 
   final Widget child;
+  final Widget body;
   final bool forceStatusIconLight;
 
   @override
   Widget build(BuildContext context) {
-    final AppTheme theme = context.theme();
+    final AppTheme theme = context.appTheme();
     final SystemUiOverlayStyle uiOverlayStyle = forceStatusIconLight == null
         ? (theme.isDark
             ? SystemUiOverlayStyle.dark
@@ -24,7 +26,7 @@ class PAppBarTransparency extends StatelessWidget {
     return PMaterial(
       child: AnnotatedRegion<SystemUiOverlayStyle>(
         value: uiOverlayStyle.copyWith(statusBarColor: Colors.transparent),
-        child: Scaffold(
+        child: body ?? Scaffold(
           backgroundColor: theme.backgroundColor ?? Colors.transparent,
           body: child,
         ),